Parseing
Added a parser. I need the input standadized(made all caps), sanatized
(removing extra spaces), and index of the’,’which I will treat the same
as a * in with the operator.I need to make and test addition so I have
to be able to parse for addition as well. I included a reverse since
upon research I found that this will make addition much easier.